Added the files needed for the admin interface, they dont show anything useful though.
Fixed some stuff in the mysqlAdmin file which when i make changes to the mysql file we should see a tiny bit better performance and lower memory usage (this will be in the near future when I finish most of the admin stuff).  

Fixed the cookies, they work now (for some reason using all the cookie variables used for setcookie lead to the cookie not being set, i'll read more about it sometime and try to get it working).

Fixed a few bugs with the feedstockAdmin file.

BIGGEST thing was login's work now for the admin interface but you don't see anything cool as of yet.